backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlmodel==0.0.16
sqlalchemy==2.0.23
alembic==1.12.1

# SQLite (开发环境)
aiosqlite==0.19.0

# PostgreSQL (生产环境)
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# 环境变量管理
python-dotenv==1.0.0

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 文件处理
python-multipart==0.0.6
python-docx==1.1.0

# 日志
loguru==0.7.2

# 开发工具
ruff==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM 集成 (Deepseek API)
openai==0.28.1

# 缓存
redis==5.0.1

# Stage 3: 深度质量分析依赖
jieba==0.42.1                      # 中文分词
scikit-learn==1.6.1               # 机器学习和文本相似度
numpy==2.2.6                     # 数值计算
matplotlib==3.10.3                 # 图表生成
seaborn==0.13.2                   # 统计图表
reportlab==4.4.1                  # PDF报告生成
jinja2==3.1.2                     # HTML模板引擎
#wordcloud==1.9.2                  # 词云生成
textstat==0.7.7                   # 文本可读性统计
supabase==2.3.4
pandas==2.2.3

# 测试依赖
pytest-mock==3.14.1
